Mark ExecutablePackageObtainer to internal
