+++ src/dotnet/ExecutablePackageObtainer/ToolConfigurationDeserialization/DotNetCliTool.cs
@@ -1,0 +1,13 @@
+﻿using System.Diagnostics;
+using System.Xml.Serialization;
+
+namespace Microsoft.DotNet.ExecutablePackageObtainer.ToolConfigurationDeserialization
+{
+    [DebuggerStepThrough]
+    [XmlRoot(Namespace = "", IsNullable = false)]
+    public class DotNetCliTool
+    {
+        [XmlArrayItem("Command", IsNullable = false)]
+        public DotNetCliToolCommand[] Commands { get; set; }
+    }
+}
