+++ src/dotnet/ExecutablePackageObtainer/ToolConfigurationDeserialization/DotNetCliToolCommand.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Diagnostics;
+using System.Xml.Serialization;
+
+namespace Microsoft.DotNet.ExecutablePackageObtainer.ToolConfigurationDeserialization
+{
+    [Serializable]
+    [DebuggerStepThrough]
+    [XmlType(AnonymousType = true)]
+    public class DotNetCliToolCommand
+    {
+        [XmlAttribute]
+        public string Name { get; set; }
+
+        [XmlAttribute]
+        public string EntryPoint { get; set; }
+
+        [XmlAttribute]
+        public string Runner { get; set; }
+    }
+}
